[QFEED-123] Refactor: 추천 기능 애그리거트 follow 아래로 변경
