+++ module-domain/src/main/java/com/wsws/moduledomain/follow/vo/UserRecommendation.java
@@ -1,4 +1,4 @@
-package com.wsws.moduledomain.recommendation;
+package com.wsws.moduledomain.follow.vo;
 
 import lombok.Value;
